chore: update HomeTabsList location
