+++ src/frontend/Navigation/Tab/index.tsx
@@ -1,0 +1,6 @@
+export type HomeTabsList = {
+  Map: undefined;
+  Camera: undefined;
+  Tracking: undefined;
+  ObservationsList: undefined;
+};
