+++ src/frontend/constants.ts
@@ -1,5 +1,5 @@
 import {AppStackList} from './Navigation/AppStack';
-import {HomeTabsList} from './Navigation/Stack/AppScreens';
+import {type HomeTabsList} from './Navigation/Tab';
 
 // this has to be a string because js does not recognize 00000 as being 5 digits
 export const OBSCURE_PASSCODE = '00000';
