+++ src/frontend/sharedTypes.ts
@@ -6,7 +6,7 @@ import {NativeStackScreenProps} from '@react-navigation/native-stack';
 
 import {MessageDescriptor} from 'react-intl';
 import {AppStackList} from './Navigation/AppStack';
-import {HomeTabsList} from './Navigation/Stack/AppScreens';
+import {type HomeTabsList} from './Navigation//Tab';
 import {Observation, ObservationValue} from '@mapeo/schema';
 import type {RoleId, RoleIdForNewInvite} from '@mapeo/core/dist/roles';
